Added Accuracy in AppSettings for Analytic View decimal places. Changed AppSettings registration in Bootstrapper to use in CutterAccessLayer to register. Commented AnalyseService.

+++ SealWatch.Code/CutterLayer/CutterAnalyseDto.cs
@@ -12,7 +12,7 @@ public class AnalysedCutterDto : BaseListDto
 
     public DateTime MillingStop { get; set; }
 
-    public int DaysLeft { get; set; }
+    public double DaysLeft { get; set; }
 
     public double Durability { get; set; }
